emea private assets to showpad

+++ src/com/depuysynthes/scripts/showpad/ShowpadMediaBinDecorator.java
@@ -15,7 +15,6 @@ import java.util.Set;
 import com.depuysynthes.scripts.DSMediaBinImporterV2;
 import com.depuysynthes.scripts.MediaBinDeltaVO;
 import com.depuysynthes.scripts.MediaBinDeltaVO.State;
-import com.siliconmtn.io.FileType;
 import com.siliconmtn.security.OAuth2TokenViaCLI;
 import com.siliconmtn.security.OAuth2TokenViaCLI.Config;
 import com.siliconmtn.util.Convert;
@@ -35,7 +34,7 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 
 	protected ShowpadApiUtil showpadApi;
 	protected List<ShowpadDivisionUtil> divisions = new ArrayList<>();
-	private boolean deduplicate = false; //override via args[1];
+	private boolean deduplicate = false; //override via args[1]
 
 	/**
 	 * @param args
@@ -45,17 +44,16 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 		super(args);
 
 		//setup the oAuth util now that the config file has been loaded
-		showpadApi = new ShowpadApiUtil(new OAuth2TokenViaCLI(new HashMap<Config, String>(){
-			private static final long serialVersionUID = -8625615784451892590L;
-			{
-				put(Config.USER_ID, props.getProperty("showpadAcctName"));
-				put(Config.API_KEY, props.getProperty("showpadApiKey"));
-				put(Config.API_SECRET, props.getProperty("showpadApiSecret"));
-				put(Config.TOKEN_CALLBACK_URL, props.getProperty("showpadCallbackUrl"));
-				put(Config.TOKEN_SERVER_URL, props.getProperty("showpadTokenUrl"));
-				put(Config.AUTH_SERVER_URL,  props.getProperty("showpadAuthUrl"));
-				put(Config.KEYSTORE, "showpad");
-			}}, Arrays.asList(props.getProperty("showpadScopes").split(","))));
+		Map<Config, String> config = new HashMap<>();
+		config.put(Config.USER_ID, props.getProperty("showpadAcctName"));
+		config.put(Config.API_KEY, props.getProperty("showpadApiKey"));
+		config.put(Config.API_SECRET, props.getProperty("showpadApiSecret"));
+		config.put(Config.TOKEN_CALLBACK_URL, props.getProperty("showpadCallbackUrl"));
+		config.put(Config.TOKEN_SERVER_URL, props.getProperty("showpadTokenUrl"));
+		config.put(Config.AUTH_SERVER_URL,  props.getProperty("showpadAuthUrl"));
+		config.put(Config.KEYSTORE, "showpad");
+		List<String> scopes = Arrays.asList(props.getProperty("showpadScopes").split(","));
+		showpadApi = new ShowpadApiUtil(new OAuth2TokenViaCLI(config, scopes));
 
 		if (args.length > 1)
 			deduplicate = Convert.formatBoolean(args[1]); //args[0] passes 'type' to the superclass, so we'll use args[1] here
@@ -84,24 +82,22 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 		//used only for de-duplication, which is more crisis-cleanup than something we do regularly.
 		if (deduplicate) {
 			Map<String, MediaBinDeltaVO> records = loadManifest();
-			Set<String> assetNames = new HashSet<>(records.size());
 			Set<String> localShowpadIds = new HashSet<>(records.size());
-			
+
 			for (MediaBinDeltaVO vo : records.values()) {
-				assetNames.add(ShowpadDivisionUtil.makeShowpadAssetName(vo, new FileType(vo.getFileNm())));
 				if (vo.getShowpadId() != null) localShowpadIds.add(vo.getShowpadId());
 			}
-			
+
 			try {
 				for (ShowpadDivisionUtil util : divisions)
-					util.cleanupShowpadDups(assetNames, localShowpadIds);
+					util.cleanupShowpadDups(localShowpadIds);
 			} catch (QuotaException qe) {
 				log.error(qe);
 			}
 			log.info("deuplication complete");
 			return;
 		}
-		
+
 		super.run();
 	}
 
@@ -112,7 +108,7 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 	 * If we try to add a tag to an asset without using it's ID, and it already existing in the system, it will fail.
 	 * @throws QuotaException 
 	 */
-	private void loadShowpadDivisionList() throws QuotaException {
+	protected void loadShowpadDivisionList() throws QuotaException {
 		String[] divs = props.getProperty("showpadDivisions").split(",");
 		for (String d : divs) {
 			String[] div = d.split("=");
@@ -152,11 +148,11 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 		log.info("loaded " + divisionAssets.size() + " divisions from the database");
 
 		//marry the divisionAssets to their respective util object
-		for (String divId : divisionAssets.keySet()) {
+		for (Map.Entry<String, Map<String, String>> entry : divisionAssets.entrySet()) {
 			for (ShowpadDivisionUtil util : divisions) {
-				if (util.getDivisionId().equals(divId)) {
-					util.setDivisionAssets(divisionAssets.get(divId));
-					log.info("gave " + divisionAssets.get(divId).size() + " assets to division=" + divId);
+				if (util.getDivisionId().equals(entry.getKey())) {
+					util.setDivisionAssets(entry.getValue());
+					log.info("gave " + entry.getValue().size() + " assets to division=" + entry.getKey());
 					break;
 				}
 			}
@@ -182,22 +178,17 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 		if (!isInsert) return;
 
 		//push all changes to Showpad
-		outer: for (MediaBinDeltaVO vo : masterRecords.values()) {
+		for (MediaBinDeltaVO vo : masterRecords.values()) {
 			//we need to sort out what gets pushed to Showpad on our own.
 			//if it's failed, being deleted, or unchanged and already in Showpad, skip it.
 			State s = vo.getRecordState();
 			if (s == State.Failed || s == State.Delete)
 				continue;
 
-			for (ShowpadDivisionUtil util : divisions) {
-				try {
-					util.pushAsset(vo);					
-				} catch (QuotaException qe) {
-					String msg = makeMessage(vo, "Could not push file to showpad: " + qe.getMessage());
-					failures.add(new Exception(msg));
-					log.error("could not push to showpad, quota reached", qe);
-					break outer;
-				}
+			try {
+				loopFileThroughDivisions(vo);
+			} catch (QuotaException e) {
+				return; //we can't continue if the quota has been reached.
 			}
 			log.info("completed: " + vo.getFileNm());
 		}
@@ -217,6 +208,25 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 	}
 
 
+	/**
+	 * called for each mediabin asset in the stack - push it out to each of the Divisions we're managing
+	 * @param vo
+	 * @throws QuotaException
+	 */
+	protected void loopFileThroughDivisions(MediaBinDeltaVO vo) throws QuotaException {
+		for (ShowpadDivisionUtil util : divisions) {
+			try {
+				util.pushAsset(vo);					
+			} catch (QuotaException qe) {
+				String msg = makeMessage(vo, "Could not push file to showpad: " + qe.getMessage());
+				failures.add(new Exception(msg));
+				log.error("could not push to showpad, quota reached", qe);
+				throw qe;
+			}
+		}
+	}
+
+
 
 	/**
 	 * override the deleteRecords methods to push deletions to Showpad after
@@ -229,46 +239,55 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 		//confirm we have something to delete
 		if (getDataCount("deleted") == 0) return;
 
-		int cnt = 0;
+		List<String> deletedIds = deleteFromShowpad(masterRecords);
+
+		//fail-fast if there's nothing to do
+		if (deletedIds == null || deletedIds.isEmpty()) return;
+
 		// Build the SQL Statement
 		StringBuilder sql = new StringBuilder(350);
 		sql.append("delete from ").append(props.getProperty(Constants.CUSTOM_DB_SCHEMA)).append("dpy_syn_showpad ");
-		sql.append("where dpy_syn_mediabin_id in ('~'");
-		for (MediaBinDeltaVO vo : masterRecords.values()) {
-			if (vo.getRecordState() == State.Delete) {
-				sql.append(",?");
-				++cnt;
-			}
-		}
-		sql.append(")");
+		sql.append("where dpy_syn_mediabin_id=?");
 		log.debug(sql);
+		try (PreparedStatement ps  = dbConn.prepareStatement(sql.toString())) {
+			for (String id : deletedIds) {
+				ps.setString(1, id);
+				ps.addBatch();
+			}
+			ps.executeBatch();
+		} catch (SQLException sqle) {
+			failures.add(sqle);
+		}
+	}
 
-		if (cnt > 0) { //don't run the query if we don't need to
-			cnt = 1;
-			try (PreparedStatement ps  = dbConn.prepareStatement(sql.toString())) {
-				for (MediaBinDeltaVO vo : masterRecords.values()) {
-					if (vo.getRecordState() == State.Delete)  {
-						try {
-							//push deletions to Showpad
-							for (ShowpadDivisionUtil util : divisions)
-								util.deleteAsset(vo);
 
-							//if success, delete it from the DB as well
-							ps.setString(cnt++, vo.getDpySynMediaBinId());
+	/**
+	 * deletes all the 'deleted' records from Showpad.
+	 * Returns a list of IDs that were deleted, so we can delete them from the local database as well.
+	 * @param masterRecords
+	 * @return
+	 * @throws QuotaException
+	 */
+	protected List<String> deleteFromShowpad(Map<String, MediaBinDeltaVO> masterRecords) {
+		List<String> data = new ArrayList<>(100);
+		for (MediaBinDeltaVO vo : masterRecords.values()) {
+			if (vo.getRecordState() != State.Delete)  continue;
 
-						} catch (QuotaException qe) {
-							String msg = makeMessage(vo, "Could not delete file from showpad: " + qe.getMessage());
-							failures.add(new Exception(msg));
-							log.error("could not delete from showpad", qe);
-							break;
-						}
-					}
-				}
-				cnt = ps.executeUpdate();
-			} catch (SQLException sqle) {
-				failures.add(sqle);
+			try {
+				//push deletions to Showpad
+				for (ShowpadDivisionUtil util : divisions)
+					util.deleteAsset(vo);
+
+				//if success, delete it from the DB as well
+				data.add(vo.getDpySynMediaBinId());
+
+			} catch (QuotaException qe) {
+				String msg = makeMessage(vo, "Could not delete asset from showpad: " + qe.getMessage());
+				failures.add(new Exception(msg));
+				log.error("could not delete from showpad", qe);
 			}
 		}
+		return data;
 	}
 
 
@@ -280,7 +299,7 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 	@Override
 	protected void countDBRecords() {
 		super.countDBRecords();
-		
+
 		StringBuilder sql = new StringBuilder(150);
 		sql.append("select count(*), division_id, case when asset_id='FAILED_PROCESSING' then 1 else 0 end as status from ");
 		sql.append(props.get(Constants.CUSTOM_DB_SCHEMA)).append("dpy_syn_showpad ");
@@ -289,18 +308,26 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 
 		try (PreparedStatement ps = dbConn.prepareStatement(sql.toString())) {
 			ResultSet rs = ps.executeQuery();
-			while (rs.next()) {
-				for (ShowpadDivisionUtil util : divisions) {
-					if (rs.getString(2).equals(util.getDivisionId()) && rs.getInt("status") == 0) {
-						util.setDbCount(rs.getInt(1));
-					} else if (rs.getString(2).equals(util.getDivisionId()) && rs.getInt("status") == 1) {
-						util.setFailCount(rs.getInt(1));
-					}
-				}
-			}
-
+			while (rs.next())
+				setRecordCounts(rs);
 		} catch (SQLException sqle) {
 			log.error("could not count records", sqle);
+		}
+	}
+
+
+	/**
+	 * sets the counter values (to be reported in the email) into each of the division utils. 
+	 * @param rs
+	 * @throws SQLException
+	 */
+	protected void setRecordCounts(ResultSet rs) throws SQLException {
+		for (ShowpadDivisionUtil util : divisions) {
+			if (rs.getString(2).equals(util.getDivisionId()) && rs.getInt("status") == 0) {
+				util.setDbCount(rs.getInt(1));
+			} else if (rs.getString(2).equals(util.getDivisionId()) && rs.getInt("status") == 1) {
+				util.setFailCount(rs.getInt(1));
+			}
 		}
 	}
 
@@ -330,7 +357,7 @@ public class ShowpadMediaBinDecorator extends DSMediaBinImporterV2 {
 					log.warn(failures.get(i).getMessage());
 				}
 			}
-			
+
 			html.append("<hr/>\r\n");
 		}
 	}
